# Conflicts:
#	.github/workflows/release-preview.yaml
#	controlplane/package.json
#	router/go.sum
#	router/pkg/app/app.go
#	router/pkg/config/config.go
#	router/pkg/graphiql/playgroundhandler.go
#	router/pkg/graphiql/playgroundhandler_test.go
#	router/pkg/metric/handler.go

+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -56,6 +56,7 @@ import { TraceRepository } from '../repositories/analytics/TraceRepository.js';
 import type { RouterOptions } from '../routes.js';
 import { ApiKeyGenerator } from '../services/ApiGenerator.js';
 import { handleError, isValidLabelMatchers, isValidLabels } from '../util.js';
+import ApolloMigrator from '../services/ApolloMigrator.js';
 
 export default function (opts: RouterOptions): Partial<ServiceImpl<typeof PlatformService>> {
   return {
@@ -1615,6 +1616,20 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           };
         }
 
+        const userRoles = await orgRepo.getOrganizationMemberRoles({
+          userID: authContext.userId || '',
+          organizationID: authContext.organizationId,
+        });
+
+        if (!(apiKey.creatorUserID === authContext.userId || userRoles.includes('admin'))) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR,
+              details: `You are not authorized to delete the api key '${apiKey.name}'`,
+            },
+          };
+        }
+
         await orgRepo.removeAPIKey({
           name: req.name,
           organizationID: authContext.organizationId,
@@ -1699,6 +1714,85 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
         });
         // deleting the user from the db
         await userRepo.deleteUser({ id: user.id });
+
+        return {
+          response: {
+            code: EnumStatusCode.OK,
+          },
+        };
+      });
+    },
+
+    migrateFromApollo: (req, ctx) => {
+      const logger = opts.logger.child({
+        service: ctx.service.typeName,
+        method: ctx.method.name,
+      });
+
+      return handleError<PlainMessage<DeleteAPIKeyResponse>>(logger, async () => {
+        const authContext = await opts.authenticator.authenticate(ctx.requestHeader);
+        const orgRepo = new OrganizationRepository(opts.db);
+        const fedGraphRepo = new FederatedGraphRepository(opts.db, authContext.organizationId);
+        const subgraphRepo = new SubgraphRepository(opts.db, authContext.organizationId);
+
+        const org = await orgRepo.byId(authContext.organizationId);
+        if (!org) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_NOT_FOUND,
+              details: `Organization not found`,
+            },
+          };
+        }
+
+        const apolloMigrator = new ApolloMigrator({ apiKey: req.apiKey, organizationSlug: org.slug });
+
+        const graph = await apolloMigrator.fetchGraphID();
+        const graphDetails = await apolloMigrator.fetchGraphDetails({ graphID: graph.id, variantName: 'main' });
+
+        if (await fedGraphRepo.exists(graph.name)) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_ALREADY_EXISTS,
+              details: `Federated graph '${graph.name}' already exists.`,
+            },
+          };
+        }
+
+        for (const subgraph of graphDetails.subgraphs) {
+          if (await subgraphRepo.exists(subgraph.name)) {
+            return {
+              response: {
+                code: EnumStatusCode.ERR_ALREADY_EXISTS,
+                details: `Subgraph '${subgraph.name}' already exists`,
+              },
+            };
+          }
+        }
+
+        const federatedGraph = await apolloMigrator.migrateGraphFromApollo({
+          fedGraph: {
+            name: graph.name,
+            routingURL: graphDetails.fedGraphRoutingURL,
+          },
+          subgraphs: graphDetails.subgraphs,
+          organizationID: authContext.organizationId,
+          db: opts.db,
+        });
+
+        const compositionErrors = await updateComposedSchema({
+          federatedGraph,
+          fedGraphRepo,
+          subgraphRepo,
+        });
+
+        if (compositionErrors.length > 0) {
+          return {
+            response: {
+              code: EnumStatusCode.ERR_SUBGRAPH_COMPOSITION_FAILED,
+            },
+          };
+        }
 
         return {
           response: {
